+++ cvat/apps/engine/task.py
@@ -40,8 +40,8 @@ from cvat.apps.engine.utils import (
     av_scan_paths, format_list,get_rq_job_meta,
     define_dependent_job, get_rq_lock_by_user, load_image
 )
-from cvat.apps.engine.quality_control import HoneypotFrameSelector
 from cvat.apps.engine.rq_job_handler import RQId
+from cvat.apps.engine.task_validation import HoneypotFrameSelector
 from cvat.utils.http import make_requests_session, PROXIES_FOR_UNTRUSTED_URLS
 from utils.dataset_manifest import ImageManifestManager, VideoManifestManager, is_manifest
 from utils.dataset_manifest.core import VideoManifestValidator, is_dataset_manifest
